web/requirements.txt
Django
django-environ==0.10.0
psycopg2-binary==2.9.9
google-cloud-secret-manager==2.16.1
django-cors-headers
dj-stripe
djangorestframework

torch
numpy
opencv-python
matplotlib
scipy
tensorflow
ultralytics
lapx